+++ src/main/kotlin/dev/stillya/vpet/graphics/Effect.kt
@@ -1,0 +1,7 @@
+package dev.stillya.vpet.graphics
+
+import java.awt.Graphics2D
+
+interface Effect {
+	fun apply(g: Graphics2D)
+}